+++ dflib/src/main/java/com/nhl/dflib/range/Range.java
@@ -3,13 +3,16 @@ package com.nhl.dflib.range;
 /**
  * @since 0.6
  */
-// TODO: start using ranges within Series, Indexes, etc.
 public class Range {
 
     public static void checkRange(int from, int len, int masterLen) {
-        // TODO: allow negative ranges to reference range from the tail?
+
         if (from < 0) {
             throw new IllegalArgumentException("Negative 'from' index: " + from);
+        }
+
+        if (len < 0) {
+            throw new IllegalArgumentException("Negative 'len': " + len);
         }
 
         if (from > masterLen) {