+++ dflib/src/main/java/com/nhl/dflib/series/DoubleArraySeries.java
@@ -52,11 +52,21 @@ public class DoubleArraySeries extends DoubleBaseSeries {
 
     @Override
     public DoubleSeries headDouble(int len) {
+
+        if (len < 0) {
+            return tailDouble(size + len);
+        }
+
         return len < size ? new DoubleArraySeries(data, offset, len) : this;
     }
 
     @Override
     public DoubleSeries tailDouble(int len) {
+
+        if (len < 0) {
+            return headDouble(size + len);
+        }
+
         return len < size ? new DoubleArraySeries(data, offset + size - len, len) : this;
     }