+++ dflib/src/test/java/com/nhl/dflib/Series_HeadTest.java
@@ -26,4 +26,11 @@ public class Series_HeadTest {
         Series<String> s = type.createSeries("a", "b", "c").head(4);
         new SeriesAsserts(s).expectData("a", "b", "c");
     }
+
+    @ParameterizedTest
+    @EnumSource(SeriesType.class)
+    public void testNegative(SeriesType type) {
+        Series<String> s = type.createSeries("a", "b", "c").head(-2);
+        new SeriesAsserts(s).expectData("c");
+    }
 }