+++ dflib/src/test/java/com/nhl/dflib/Series_TailTest.java
@@ -26,4 +26,11 @@ public class Series_TailTest {
         Series<String> s = type.createSeries("a", "b", "c").tail(4);
         new SeriesAsserts(s).expectData("a", "b", "c");
     }
+
+    @ParameterizedTest
+    @EnumSource(SeriesType.class)
+    public void testNegative(SeriesType type) {
+        Series<String> s = type.createSeries("a", "b", "c").tail(-2);
+        new SeriesAsserts(s).expectData("a");
+    }
 }