+++ dflib/src/test/java/com/nhl/dflib/series/IntSequenceSeriesTest.java
@@ -13,5 +13,14 @@ public class IntSequenceSeriesTest {
 
         IntSequenceSeries s2 = new IntSequenceSeries(1, 5);
         assertEquals(4, s2.size());
+
+        IntSequenceSeries s3 = new IntSequenceSeries(0, 1);
+        assertEquals(1, s3.size());
+    }
+
+    @Test
+    public void testSizeZero() {
+        IntSequenceSeries s1 = new IntSequenceSeries(0, 0);
+        assertEquals(0, s1.size());
     }
 }