+++ dflib/src/test/java/com/nhl/dflib/series/IntSequenceSeries_TailIntTest.java
@@ -1,6 +1,6 @@
-package com.nhl.dflib;
+package com.nhl.dflib.series;
 
-import com.nhl.dflib.series.IntSequenceSeries;
+import com.nhl.dflib.IntSeries;
 import com.nhl.dflib.unit.IntSeriesAsserts;
 import org.junit.jupiter.api.Test;
 
@@ -28,5 +28,11 @@ public class IntSequenceSeries_TailIntTest {
     public void test_OutOfBounds() {
         IntSeries s = new IntSequenceSeries(1, 4).tailInt(4);
         new IntSeriesAsserts(s).expectData(1, 2, 3);
+    }
+
+    @Test
+    public void test_Negative() {
+        IntSeries s = new IntSequenceSeries(1, 4).tailInt(-2);
+        new IntSeriesAsserts(s).expectData(1);
     }
 }